+++ R/style_tt.R
@@ -27,7 +27,6 @@
 #'   - Color names with luminance levels from [the `ninecolors` package](https://mirror.quantum5.ca/CTAN/macros/latex/contrib/ninecolors/ninecolors.pdf) (ex: "azure4", "magenta8", "teal2", "gray1", "olive3"). 
 #' @param background Background color. Specified as a color name or hexadecimal code. Can be `NULL` for default color.
 #' @param fontsize Font size in em units. Can be `NULL` for default size.
-#' @param width Width of column in em units. Can be `NULL` for default width.
 #' @param align A single character or a string with a number of characters equal to the number of columns in `j`. Valid characters include 'c' (center), 'l' (left), 'r' (right), 'd' (decimal). Decimal alignment is only available in LaTeX via the `siunitx` package. The width of columns is determined by the maximum number of digits to the left and to the right in all cells specified by `i` and `j`.
 #' @param alignv A single character specifying vertical alignment. Valid characters include 't' (top), 'm' (middle), 'b' (bottom).
 #' @param colspan Number of columns a cell should span. `i` and `j` must be of length 1.
@@ -131,7 +130,6 @@ style_tt <- function (x,
                       color = NULL,
                       background = NULL,
                       fontsize = NULL,
-                      width = NULL,
                       align = NULL,
                       alignv = NULL,
                       colspan = NULL,
@@ -150,6 +148,10 @@ style_tt <- function (x,
 
   out <- x
 
+  if ("width" %in% names(list(...))) {
+    stop("The `width` argument is now in the `tt()` function.", call. = FALSE)
+  }
+
   # issue #759: reuse object with different styles across RevealJS slides requires new ID every time style_tt is called
   out@id <- get_id("tinytable_")
 
@@ -166,7 +168,6 @@ style_tt <- function (x,
               color = color,
               background = background,
               fontsize = fontsize,
-              width = width,
               align = align,
               alignv = alignv,
               colspan = colspan,
@@ -208,7 +209,6 @@ style_tt_lazy <- function (x,
                            color,
                            background,
                            fontsize,
-                           width,
                            align,
                            alignv,
                            colspan,
@@ -242,13 +242,17 @@ style_tt_lazy <- function (x,
         assert_choice(align_character, c("c", "l", "r", "d"))
       }
       if (is.null(j)) {
-        msg <- "Please specify the `j` argument."
-        stop(msg, call. = FALSE)
+        j <- seq_len(x@ncol)
       }
     }
   }
-    
-  nalign <- if (is.null(j)) x@ncol else length(j)
+
+  if (x@output == "typst") {
+    nalign <- x@ncol
+  } else {
+    nalign <- if (is.null(j)) x@ncol else length(j)
+  }
+
   if (!is.null(align)) {
     align <- strsplit(align, split = "")[[1]]
     if (length(align) != 1 && length(align) != nalign) {
@@ -268,18 +272,14 @@ style_tt_lazy <- function (x,
 
   assert_style_tt(
     x = out, i = i, j = j, bold = bold, italic = italic, monospace = monospace, underline = underline, strikeout = strikeout,
-    color = color, background = background, fontsize = fontsize, width = width, align = align, alignv = alignv, 
+    color = color, background = background, fontsize = fontsize, align = align, alignv = alignv, 
     colspan = colspan, rowspan = rowspan, indent = indent,
     line = line, line_color = line_color, line_width = line_width,
     tabularray_inner = tabularray_inner, tabularray_outer = tabularray_outer, bootstrap_css = bootstrap_css,
     bootstrap_css_rule = bootstrap_css_rule, bootstrap_class = bootstrap_class)
 
-  if (!is.null(width)) {
-    width <- paste0(width, "em")
-  }
 
-
-  out <- style_eval(x = out, i = i, j = j, bold = bold, italic = italic, monospace = monospace, underline = underline, strikeout = strikeout, color = color, background = background, fontsize = fontsize, width = width, align = align, alignv = alignv, colspan = colspan, rowspan = rowspan, indent = indent, tabularray_inner = tabularray_inner, tabularray_outer = tabularray_outer, bootstrap_css = bootstrap_css, bootstrap_css_rule = bootstrap_css_rule, bootstrap_class = bootstrap_class, line = line, line_color = line_color, line_width = line_width)
+  out <- style_eval(x = out, i = i, j = j, bold = bold, italic = italic, monospace = monospace, underline = underline, strikeout = strikeout, color = color, background = background, fontsize = fontsize, align = align, alignv = alignv, colspan = colspan, rowspan = rowspan, indent = indent, tabularray_inner = tabularray_inner, tabularray_outer = tabularray_outer, bootstrap_css = bootstrap_css, bootstrap_css_rule = bootstrap_css_rule, bootstrap_class = bootstrap_class, line = line, line_color = line_color, line_width = line_width)
 
   return(out)
 }
@@ -296,7 +296,6 @@ assert_style_tt <- function (x,
                              color,
                              background,
                              fontsize,
-                             width,
                              align,
                              alignv,
                              colspan,
@@ -311,14 +310,8 @@ assert_style_tt <- function (x,
                              bootstrap_css = NULL,
                              bootstrap_css_rule = NULL) {
 
-  if (!is.null(width) && !is.null(i)) {
-    msg <- "The `width` argument cannot be used with `i`."
-    stop(msg, call. = FALSE)
-  }
-
   assert_integerish(colspan, len = 1, lower = 2, null.ok = TRUE)
   assert_integerish(rowspan, len = 1, lower = 2, null.ok = TRUE)
-  assert_numeric(width, len = 1, lower = 0, null.ok = TRUE)
   assert_numeric(indent, len = 1, lower = 0)
   assert_character(background, null.ok = TRUE)
   assert_character(color, null.ok = TRUE)